pyproject.toml
[project]
name = "mups"
version = "0.1.0"
description = "Mojo Ultimate Package Standard"
authors = [
    {name = "Drunkwcodes", email = "drunkwcodes@gmail.com"},
]
dependencies = [
    "pyserde>=0.12.2",
    "packaging>=23.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.3.1",
    "pylint>=2.17.4",
    "black[jupyter]>=23.7.0",
    "ipython>=8.14.0",
    "requests>=2.31.0",
    "requests-toolbelt>=1.0.0",
    "towncrier>=23.6.0",
]

[tool.towncrier]
package = "mups"
package_dir = "src"
name = "MUPS"
directory = "news"
filename = "CHANGELOG.md"